requirements.txt
fastapi>=0.109.2
google-genai==1.18.0
uvicorn>=0.15.0,<0.22.0
python-dotenv>=0.19.0
mangum>=0.17.0
python-multipart>=0.0.5
pydantic==2.6.1